requirements.txt
# Automatically

absl-py==2.3.0
accelerate==1.10.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.34.74
botocore==1.34.162
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chumpy==0.70
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.0.4
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
diskcache==5.6.3
docstring_parser==0.16
e3nn==0.4.0
einops==0.8.1
embreex==2.17.7.post6
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gdown==5.2.0
gradio==5.17.1
gradio_client==1.7.1
grpcio==1.73.0

# Updated 3D Vision
gsplat==1.5.3
pytorch3d==0.7.8
open3d==0.19.0
trimesh==4.6.12
pycolmap==3.11.1

# SAM2 for segmentation
# Install from source: https://github.com/facebookresearch/segment-anything-2
# git clone https://github.com/facebookresearch/segment-anything-2.git && cd segment-anything-2 && pip install -e .

h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.2.29
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4

lightning-utilities==0.14.3
lpips==0.1.4
lxml==5.4.0
manifold3d==3.1.1
mapbox_earcut==1.0.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
seaborn==0.13.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
narwhals==1.44.0

nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nerfview==0.0.2
nest-asyncio==1.6.0
netron==8.5.1
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
notebook==7.4.4
notebook_shim==0.2.4

# Updated Scientific Computing
numpy==1.24.4
scipy==1.9.1
pandas==2.3.0
scikit-learn==1.7.0
scikit-image==0.20.0

nuscenes-devkit==1.1.9
rosbags==0.11.0

# Updated Torch stack
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
pytorch-lightning==2.5.2
xformers==0.0.25.post1
triton==3.1.0
torchmetrics==1.8.2

opt_einsum==3.4.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.17.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
plyfile==1.1.2
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyceres==2.4
pycocotools==2.0.10
pycollada==0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
rtree==1.4.0
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
Send2Trash==1.8.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
typeguard==2.13.3
typer==0.16.0
typing_extensions==4.14.0
tyro==0.9.1
urllib3==2.5.0
uvicorn==0.34.3
viser==0.2.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xarray==2025.6.1
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0

# Optimization
torch-optimi==0.3.3

# NVIDIA CUDA Libraries
nvidia-cublas-cu12==12.1.3.1
nvidia-cudnn-cu12==9.1.0.70
nvidia-nccl-cu12==2.21.5
nvidia-cuda-runtime-cu12==12.1.105